Code cleanup and restructure controllers package
