+++ internal/router/router.go
@@ -1,7 +1,7 @@
 package router
 
 import (
-	"github.com/NodeFactoryIo/vedran/internal/controlers"
+	"github.com/NodeFactoryIo/vedran/internal/controllers"
 	"github.com/NodeFactoryIo/vedran/internal/repositories"
 	"github.com/asdine/storm/v3"
 	"github.com/gorilla/mux"
@@ -12,9 +12,9 @@ func CreateNewApiRouter(db *storm.DB) *mux.Router {
 	// initialize repos
 	nodeRepo := repositories.NewNodeRepo(db)
 	// initialize controllers
-	baseController := controlers.NewBaseController(nodeRepo)
+	apiController := controllers.NewApiController(nodeRepo)
 	// map controllers handlers to endpoints
-	router.HandleFunc("/api/v1/nodes", baseController.RegisterHandler).Methods("POST").Name("/api/v1/nodes")
+	router.HandleFunc("/api/v1/nodes", apiController.RegisterHandler).Methods("POST").Name("/api/v1/nodes")
 
 	return router
 }
